Add video to plugin guide
I created a free tutorial video about Nuxt Plugins, and this commit adds it to the guides.  

The html I used is a modified version of what I found in `/en/guide/installation.md`.

Hopefully readers of the guides will be able to benefit from the video as-is, but if there are changes you'd like me to make, then let me know and I can do a "docs edit" based on your recommendations.